requirements.txt
pytest==5.3.5
selenium3.141.0
envparse==0.2.0
